fix: 메인 페이지 동물 카드 가로 길이 통일 및 공용 컴포넌트 분리
